IGAPP-640: Move categories route to web-ts
